requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
httpx==0.27.0
mcp==1.6.0
mcp-server==0.1.4
pydantic==2.8.0
pydantic-settings==2.9.1 